add lane level resolution to full-spec
Added a high level vehicleImpact Enum that specifies the overall effect of the work zone on vehicles.
Added a LaneImpacts list that contains laneType objects that specify the impact to individual lanes.
Changed restrictions into an object (data frame) to allow precise measures. Restrictions are now nested beneath LaneImpacts
Removed OpenLanes and ClosedLanes. This information is now contained in the LaneImpacts list